Move config file functions to cfg.py
